[refactor] request, response 변경 #22
